Refactor command line flags and constructor params
This commit refactors the flags and parameters for the command line
application, the Plack App, and the Plack Middleware. Specifically,
the `defaults` option no longer exists (its sub-options are moved
up and prefixed with `default_`), and the middleware no longer
defines all the flags supported by the App. Instead, it has one
flag called `opts` which accepts a hash-ref of options that will
be forwarded to the App.