Add support for ignore files
This commit introduces support for ignore files, in the format of
the Gitignore file, which can be used to prevent access to certain
files within the served directories. Any request that matches a rule
from an ignore file will result in a 404 Not Found response.

The default ignore file is ".mccsignore", but this can be overridden
using the `ignore_file` option. In vhost mode, every host can have
its own ignore file, plus all hosts can have a global ignore file.
Both the host-specific file and the global ignore file will be used,
if they exist.